% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_syl.R
\name{syl}
\alias{syl}
\title{TH Syllables
builds syllables around vowels and consonant clusters}
\usage{
syl(prevlet, nsyl = 1, pos = "first")
}
\arguments{
\item{prevlet}{character of previous letter}

\item{nsyl}{positive integer number of syllables}

\item{pos}{string "first" or "subseq" position of letters}
}
\value{
character "syllable" of consonants and vowels
}
\description{
TH Syllables
builds syllables around vowels and consonant clusters
}
\examples{
syl("b", nsyl=1, pos="first")
}
