% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_cons.R
\name{th_clusts}
\alias{th_clusts}
\title{TH Consonant maker
constructs permitted consonant clusters}
\usage{
th_clusts(prevlet, pos = "first")
}
\arguments{
\item{prevlet}{character of previous letter}

\item{pos}{string "first" or "subseq" position of letters}
}
\value{
single or cluster of letter characters
}
\description{
TH Consonant maker
constructs permitted consonant clusters
}
\examples{
th_clusts("b", pos="first")
}
