% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_vows.R
\name{th_vows}
\alias{th_vows}
\title{TH vowels
if a vowel is chosen, decide if standalone vowel, diphthong, or triphthong}
\usage{
th_vows(prevlet)
}
\arguments{
\item{prevlet}{character of previous letter}
}
\value{
character single or cluster of vowels
}
\description{
TH vowels
if a vowel is chosen, decide if standalone vowel, diphthong, or triphthong
}
\examples{
th_vows("a")
}
