% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_noun.R
\name{thannoun}
\alias{thannoun}
\title{TH word maker
brings other TH functions to construct final word}
\usage{
thannoun(nsyl = 1)
}
\arguments{
\item{nsyl}{integer number of syllables}
}
\value{
character "word" of chosen syllabic length
}
\description{
TH word maker
brings other TH functions to construct final word
}
\examples{
x <- for(i in 1:50){ print(thannoun(nsyl=2)) }
}
